ize is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 /**
- * main - print the sizeof various types on the computer
+ * main - Entry point
  *
  * Return: Always 0 (success)
  */
@@ -13,10 +13,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("size of a char: %d byte(s)\n", sizeof(char));
-	printf("size of a int: %d byte(s)\n", sizeof(int));
-	printf("size of a long int: %d byte(s)\n", sizeof(long int));
-	printf("size of a long long int:%d byte(s)\n", sizeof(long long int));
-	printf("size of a float: %d byte(s)\n", sizeof(float));
+	printf("size of a char: %d byte(s)\n", (unsigned long)sizeof(char));
+	printf("size of a int: %d byte(s)\n", (unsigned long)sizeof(int));
+	printf("size of a long int: %d byte(s)\n", (unsigned long)sizeof(long int));
+	printf("size of a long long int:%d byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("size of a float: %d byte(s)\n", (unsigned long)sizeof(float));
 	return0;
 }
